+++ src/com/tutorial/main/FastEnemy.java
@@ -10,7 +10,7 @@ import java.awt.Rectangle;
  */
 public class FastEnemy extends GameObject {
 
-	public FastEnemy(int x, int y, ID id,Handler handler) {
+	public FastEnemy(float x, float y, ID id,Handler handler) {
 		super(x, y, id,handler);
 		velX = 2;
 		velY = 9;
@@ -33,13 +33,13 @@ public class FastEnemy extends GameObject {
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.cyan);
-		g.fillRect(x, y,16, 16);
+		g.fillRect((int)x, (int)y,16, 16);
 		
 	}
 
 	@Override
 	public Rectangle getBounds() {
-		return new Rectangle (x,y,16,16);
+		return new Rectangle ((int)x,(int)y,16,16);
 	}
