+++ src/com/tutorial/main/GameObject.java
@@ -10,13 +10,13 @@ import java.awt.Rectangle;
  */
 public abstract class GameObject {
 	//Coordenadas en la pantalla
-	protected int x;
-	protected int y;
+	protected float x;
+	protected float y;
 	//Id del objeto de juego
 	protected ID id;
 	//Variables de velocidad
-	protected int velX;
-	protected int velY;
+	protected float velX;
+	protected float velY;
 	//Handler
 	protected Handler handler;
 	
@@ -30,7 +30,7 @@ public abstract class GameObject {
 	 */
 	
 	
-	public GameObject(int x, int y, ID id, Handler handler) {
+	public GameObject(float x, float y, ID id, Handler handler) {
 		super();
 		this.x = x;
 		this.y = y;
@@ -60,7 +60,7 @@ public abstract class GameObject {
 
 	public abstract Rectangle getBounds();
 	
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
@@ -68,7 +68,7 @@ public abstract class GameObject {
 		this.x = x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 
@@ -84,7 +84,7 @@ public abstract class GameObject {
 		this.id = id;
 	}
 
-	public int getVelX() {
+	public float getVelX() {
 		return velX;
 	}
 
@@ -92,7 +92,7 @@ public abstract class GameObject {
 		this.velX = velX;
 	}
 
-	public int getVelY() {
+	public float getVelY() {
 		return velY;
 	}
