+++ src/com/tutorial/main/HUD.java
@@ -9,14 +9,14 @@ import java.awt.Graphics;
  */
 public class HUD {
 	//Controla la vida
-	public static int HEALTH = 100;
+	public static float HEALTH = 100;
 	//el valor de verde
-	private int greenValue = 255;
+	private float greenValue = 255;
 	
 	private int score = 0;
 	private int level =1;
 	public void tick() {
-		HEALTH = Game.clamp(HEALTH, 0, 100);
+		HEALTH =  Game.clamp(HEALTH, 0, 100);
 		greenValue = HEALTH *2;
 		greenValue = Game.clamp(greenValue, 0, 255);
 		score++;
@@ -25,8 +25,8 @@ public class HUD {
 	public void render(Graphics g) {
 		g.setColor(Color.gray);
 		g.fillRect(15, 15, 200, 32);
-		g.setColor(new Color(150,greenValue,0));
-		g.fillRect(15, 15, HEALTH *2, 32);
+		g.setColor(new Color(150,(int)greenValue,0));
+		g.fillRect(15, 15,(int) HEALTH *2, 32);
 		g.setColor(Color.white);
 		g.drawRect(15, 15, 200, 32);
