+++ src/com/tutorial/main/Trail.java
@@ -14,11 +14,11 @@ public class Trail extends GameObject{
 	private float alpha = 1;
 	private Color color;
 	
-	private int width;
-	private int height;
+	private float width;
+	private float height;
 	private float life;
 
-	public Trail(int x, int y, ID id, Color color, int width, int height,float life, Handler handler) {
+	public Trail(float x, float y, ID id, Color color, float width, float height,float life, Handler handler) {
 		super(x, y, id,handler);
 		this.color = color;
 		this.width = width;
@@ -41,7 +41,7 @@ public class Trail extends GameObject{
 		Graphics2D g2d = (Graphics2D) g; 
 		g2d.setComposite(makeTransparent(alpha));
 		g.setColor(color);
-		g.fillRect(x, y, width, height);
+		g.fillRect((int)x, (int)y, (int)width, (int)height);
 		//Importante resetear la transparencia a 1
 		g2d.setComposite(makeTransparent(1));
